--- src/test/java/com/exemplo/meuprojeto/MeuProjetoApplicationTests.java
@@ -1,14 +1,0 @@
-package com.exemplo.meuprojeto;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MeuProjetoApplicationTests {
-
-	@Test
-	void contextLoads() {
-		
-	}
-
-}
